Результаты, разные тесты, персональный словарь

+++ Client/Extensions/ContextExtensions.cs
@@ -7,6 +7,8 @@ public static class ContextExtensions
 {
 	public static async Task TryDeleteMessageAsync(this ITelegramBotClient client, ChatId chatId, int? messageId)
 	{
+		if(messageId == null)
+			return;
 		try
 		{
 			await client.DeleteMessageAsync(chatId, (int)messageId);
